9/22/2020
Attempted to begin to use python to control the motor instead of java so that we can use it on more than one raspberry pi version, I am not going to continue to try this because I do no think it will work and it adds an unnecissary level of complexity and complication on my end. I still have not figured out how to get the .jar to launch, I will need to find a way to do that. Some arduino motor control stuff.